+++ src/main.ts
@@ -71,7 +71,7 @@ export default class Omd2TypstPlugin extends Plugin {
   }
 
   async loadSettings() {
-    this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData());
+    this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData() as Partial<Omd2TypstSettings>);
   }
 
   async saveSettings() {